提测，需求 600设备管理 601照明管理 614数字音响



See merge request !7

+++ src/main/java/com/hoolink/manage/base/constant/Constant.java
@@ -11,7 +11,7 @@ public class Constant {
     public static final String USER_ROLE_NAME = "系统管理员";
     public static final String INITIAL_PASSWORD = "123456";
     
-    public static final int LEVEL_ONE = 1;
-    public static final int LEVEL_TWO = 2;
-    public static final int LEVEL_THREE = 3;
+    public static final Byte LEVEL_ONE = 1;
+    public static final Byte LEVEL_TWO = 2;
+    public static final Byte LEVEL_THREE = 3;
 }
